@@ -25,6 +25,7 @@
   "homepage": "https://github.com/GNYIO/gny-experiment#readme",
   "dependencies": {
     "@nodeutils/defaults-deep": "^1.1.0",
+    "@polkadot/ts": "git+https://github.com/a1300/ts.git",
     "@types/bip39": "^2.4.0",
     "@types/body-parser": "=1.17.0",
     "@types/bs58": "^3.0.30",
@@ -95,7 +96,7 @@
     "ts-jest": "^23.1.4",
     "ts-node": "=7.0.1",
     "typeorm": "^0.2.12",
-    "typescript": "^3.0.3",
+    "typescript": "^3.3.3",
     "validate.js": "=0.12.0",
     "z-schema": "https://github.com/AschPlatform/z-schema/tarball/v3.15.2+asch.1"
   },
