@@ -44,7 +44,7 @@ const DELAY = (ms = 5000) => new Promise((resolve, reject) => setTimeout(resolve
     const proc = exec('node app.js', {
       cwd: dir,
      }, (err, stdout, stderr) => {
-       console.log(`${i} node started`);
+       console.log(`[${i}] node started; err: ${err}; stdout: ${stdout}; stderr: ${stderr}`);
 
      });
 
