@@ -2,50 +2,69 @@
 import { Bundle } from './bundle';
 import * as PeerId from 'peer-id';
 import { extractIpAndPort } from './util';
-
-import { P2PMessage } from '../../src/interfaces';
+import { P2PMessage, ILogger, P2PSubscribeHandler } from '../../src/interfaces';
 
 export class Peer2Peer {
   private bundle: Bundle;
-  private peerInfo = null;
-  private bootstrapNode: string;
+  private logger: ILogger;
 
-  constructor (peerInfo: any, bootstrapNode: string) {
-    this.peerInfo = peerInfo;
-    this.bootstrapNode = bootstrapNode;
+  public isStarted = () => {
+    return this.bundle.isStarted();
+  }
+  get peerInfo(): PeerInfo {
+    return this.bundle.peerInfo;
+  }
+  get peerBook(): PeerBook {
+    return this.bundle.peerBook;
   }
 
-  public startAsync = async () => {
-    const peerInfo = this.peerInfo;
+  constructor (logger: ILogger, peerInfo: any, bootstrapNode: string, bootStrapInterval: number = 30000) {
+    this.logger = logger;
+
     const configuration = {
       peerInfo,
       config: {
         peerDiscovery: {
           bootstrap: {
-            list: this.bootstrapNode ? [ this.bootstrapNode ] : [],
+            list: bootstrapNode ? [ bootstrapNode ] : [],
+            interval: bootStrapInterval,
           },
         },
       },
     };
 
     this.bundle = new Bundle(configuration);
+  }
+
+  public startAsync = async () => {
 
     this.bundle.on('stop', this.stopped);
     this.bundle.on('error', this.errorOccurred);
-    this.bundle.on('peer:connect', this.addPeerToDb);
-    this.bundle.on('peer:disconnect', this.removePeerFromDb);
+    this.bundle.on('peer:connect', this.peerConnect);
+    this.bundle.on('peer:disconnect', this.peerDisconnect);
     this.bundle.on('peer:discovery', this.peerDiscovery);
     await this.bundle.startAsync();
+    this.printOwnPeerInfo();
+  }
+
+  public stopAsync = async () => {
+    await this.bundle.stopAsync();
+  }
+
+  private printOwnPeerInfo = () => {
+    let addresses = '';
+    this.peerInfo.multiaddrs.forEach((adr) => addresses += `\t${adr.toString()}\n`);
+    this.logger.info(`\n[P2P] started node: ${this.peerInfo.id.toB58String()}\n${addresses}`);
   }
 
   private stopped = (err) => {
-    global.app.logger.info('[P2P] p2p node stopped');
+    this.logger.info('[P2P] p2p node stopped');
   }
 
   private errorOccurred = (err) => {
-    global.app.logger.error(`[P2P] error occurred: ${err.message}`);
+    this.logger.error(`[P2P] error occurred: ${err.message}`);
     if (typeof err.message === 'string' && err.message.includes('EADDRINUSE')) {
-      global.app.logger.warn('port is already in use, shutting down...');
+      this.logger.warn('port is already in use, shutting down...');
       throw new Error(err);
     }
   }
@@ -54,17 +73,17 @@ export class Peer2Peer {
     this.bundle.stop(cb);
   }
 
-  subscribe (topic: string, handler: any) {
+  subscribe (topic: string, handler: P2PSubscribeHandler) {
     const filterBroadcastsEventHandler = (message: P2PMessage) => {
       // this filters messages out which are published from the own node
-      if (message.from === this.bundle.peerInfo.id.toB58String()) {
+      if (message.from === this.peerInfo.id.toB58String()) {
         return;
       }
 
       const id = PeerId.createFromB58String(message.from);
       this.bundle.peerRouting.findPeer(id, {}, (err, result) => { // find peer in routing table that broadcasted message
         if (err) {
-          global.app.logger.warn('could not find peer that broadcasted message');
+          this.logger.warn('could not find peer that broadcasted message');
           return;
         }
 
@@ -76,9 +95,9 @@ export class Peer2Peer {
           handler(extendedMsg);
         };
 
-        this.bundle.dial(result, (erro, conn) => { // dial to peer that broadcasted message
+        this.bundle.dial(result, (erro) => { // dial to peer that broadcasted message
           if (erro) {
-            global.app.logger.warn(`could not dial peer ${id}`);
+            this.logger.warn(`could not dial peer ${id}`);
             return;
           }
           return finish(result);
@@ -101,7 +120,7 @@ export class Peer2Peer {
     });
   }
 
-  broadcastTransactionAsync(data: Buffer): Promise<void> {
+  public broadcastTransactionAsync(data: Buffer): Promise<void> {
     return new Promise((resolve, reject) => {
       if (!this.bundle.isStarted()) {
         resolve();
@@ -113,7 +132,7 @@ export class Peer2Peer {
     });
   }
 
-  broadcastNewBlockHeaderAsync(data: Buffer): Promise<void> {
+  public broadcastNewBlockHeaderAsync(data: Buffer): Promise<void> {
     return new Promise((resolve, reject) => {
       if (!this.bundle.isStarted()) {
         resolve();
@@ -125,37 +144,54 @@ export class Peer2Peer {
     });
   }
 
-  private addPeerToDb(peer) {
-    // TODO implement
-    global.app.logger.info(`[P2P] peer:connect:${peer.id.toB58String()}`);
+  broadcastAsync(topic: string, data: Buffer): Promise<void> {
+    return new Promise((resolve, reject) => {
+      if (!this.bundle.isStarted()) {
+        resolve();
+      }
+      this.bundle.pubsub.publish(topic, data, (err) => {
+        if (err) reject(err.message);
+        else resolve();
+      });
+    });
   }
 
-  private removePeerFromDb = (peer) => {
-    // TODO implemnet
-    global.app.logger.info(`[P2P] peer:disconnect:${peer.id.toB58String()}`);
+  private peerConnect = (peer) => {
+    this.logger.info(`[P2P] peer:connect:${peer.id.toB58String()}`);
+  }
+
+  private peerDisconnect = (peer) => {
+    this.logger.info(`[P2P] peer:disconnect:${peer.id.toB58String()}`);
     this.bundle.peerBook.remove(peer);
   }
 
-  private peerDiscovery = async (peer) => {
-    // do not spam log output: the bootstrap mechanism tries every 30s to connect to the bootstrap node(s)
+  private peerDiscovery = async (peer) => { // when peer is discovered by DHT
+    // do not spam log output: the bootstrap mechanism tries every x seconds to connect to the bootstrap node(s)
     if (!this.bundle.peerBook.has(peer)) {
-      global.app.logger.info(`[P2P] discovered peer: ${peer.id.toB58String()}`);
+      this.logger.info(`[P2P] discovered peer: ${peer.id.toB58String()}`);
+    }
+    if (!this.bundle.isStarted()) {
+      this.logger.warn(`[P2P] own node not started, can not dial to peer: ${peer.id.toB58String()}`);
+      return;
     }
     try {
       // this action establishes a __Connection__ to the newly discovered peer
       // this also adds the peer to the peerBook so the pubsub mechanism can publish to this peer
       await this.bundle.dialAsync(peer);
     } catch (err) {
-      global.app.logger.info(`[P2P] could not dial to ${peer.id.toB58String()}`);
+      this.logger.info(`[P2P] could not dial to ${peer.id.toB58String()}`);
     }
   }
 
   public getRandomNode = () => {
-    const peerInfo = this.bundle.getRandomPeer();
-    if (peerInfo) {
+    const allPeers = this.peerBook.getAllArray();
+    if (allPeers.length > 0) {
+      const index = Math.floor(Math.random() * allPeers.length);
+      const peerInfo = allPeers[index];
       const extracted = extractIpAndPort(peerInfo);
-      global.app.logger.info(`[P2P] getRandomPeer: ${peerInfo.id.toB58String()}; ${JSON.stringify(extracted)}`);
+      this.logger.info(`[P2P] getRandomPeer: ${peerInfo.id.toB58String()}; ${JSON.stringify(extracted)}`);
       return extracted;
     }
+    return undefined;
   }
 }
